Split ViewComplexTest to avoid timeouts after being moved away from 'long' junits
patch by Berenguer Blas; reviewed by Ekaterina Dimitrova for CASSANDRA-16711
