+++ test/unit/org/apache/cassandra/cql3/ViewComplexTTLTest.java
@@ -38,12 +38,13 @@ import static org.apache.cassandra.cql3.ViewComplexTest.updateView;
 import static org.apache.cassandra.cql3.ViewComplexTest.updateViewWithFlush;
 import static org.junit.Assert.assertTrue;
 
-/* This class been split into multiple ones bc of timeout issues CASSANDRA-16670
+/* ViewComplexTest class has been split into multiple ones because of timeout issues (CASSANDRA-16670)
  * Any changes here check if they apply to the other classes:
  * - ViewComplexUpdatesTest
  * - ViewComplexDeletionsTest
  * - ViewComplexTTLTest
  * - ViewComplexTest
+ * - ViewComplexLivenessTest
  */
 @RunWith(Parameterized.class)
 public class ViewComplexTTLTest extends CQLTester
