+++ test/unit/org/apache/cassandra/cql3/ViewTest.java
@@ -73,6 +73,7 @@ public class ViewTest extends CQLTester
     {
         requireNetwork();
     }
+
     @Before
     public void begin()
     {
