+++ test/unit/org/apache/cassandra/cql3/validation/operations/InsertUpdateIfConditionStaticsTest.java
@@ -29,7 +29,7 @@ import org.junit.runners.Parameterized;
 
 import org.apache.cassandra.cql3.CQLTester;
 
-/* This class been split into multiple ones bc of timeout issues CASSANDRA-16670
+/* InsertUpdateIfConditionCollectionsTest class has been split into multiple ones because of timeout issues (CASSANDRA-16670)
  * Any changes here check if they apply to the other classes
  * - InsertUpdateIfConditionStaticsTest
  * - InsertUpdateIfConditionCollectionsTest
